Fix mass assignment
Prevents the usage of write_attribute when mass assignment is supposed
to be done. In mass assignment, each attribute's setter method should
be invoked. In the case of the new method, mass assignment was already
done - so we just skip re-assigning the value.

Also, prevent invoking the provided block twice in the redefinition of
new.